# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "gpt-terminal-plus API",
    "version": "0.1.0",
    "description": "OpenAPI generated from JSDoc annotations.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Public base URL"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_TOKEN"
      }
    },
    "schemas": {
      "ExecutionResult": {
        "type": "object",
        "properties": {
          "stdout": {
            "type": "string"
          },
          "stderr": {
            "type": "string"
          },
          "exitCode": {
            "type": "integer"
          },
          "error": {
            "type": "boolean"
          }
        },
        "required": [
          "stdout",
          "stderr",
          "exitCode",
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/command/execute": {
      "post": {
        "operationId": "executeCommand",
        "summary": "Execute using first available mode",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "command": {
                    "type": "string"
                  }
                },
                "required": [
                  "command"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Execution complete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/components/schemas/ExecutionResult"
                    },
                    "aiAnalysis": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "result"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "x-openai-isConsequential": false
      }
    },
...
    "/shell/session/start": {
      "post": {
        "operationId": "shellSessionStart",
        "summary": "Start a new persistent shell session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "shell": {
                    "type": "string",
                    "default": "bash"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session started"
          },
          "500": {
            "description": "Failed to start session"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/shell/session/{id}/exec": {
      "post": {
        "operationId": "shellSessionExec",
        "summary": "Execute command inside existing session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "command": {
                    "type": "string"
                  },
                  "shell": {
                    "type": "string"
                  }
                },
                "required": [
                  "command"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command executed"
          },
          "400": {
            "description": "Command required"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/shell/session/{id}/stop": {
      "post": {
        "operationId": "shellSessionStop",
        "summary": "Stop a persistent shell session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session stopped"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Failed to stop session"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/shell/session/list": {
      "get": {
        "operationId": "shellSessionList",
        "summary": "List active shell sessions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of sessions"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/shell/session/{id}/logs": {
      "get": {
        "operationId": "shellSessionLogs",
        "summary": "Fetch logs from a shell session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session logs"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "x-openai-isConsequential": false
      }
    }
  },
...
}